pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openaireview"
version = "0.1.0"
description = "AI-powered academic paper reviewer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
]

[project.scripts]
openaireview = "reviewer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
reviewer = ["viz/*.html"]

[project.optional-dependencies]
dev = ["beautifulsoup4>=4.14.3", "lxml>=6.0.2"]
